+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed event text readability on colored backgrounds ([#1065])
 - Fixed invisible current time indicator in weekly view ([#99])
 - Fixed stuck zoom level in weekly view on some devices ([#621])
+- Fixed long-pressing the only visible event type in the quick filter not restoring all event types ([#1092])
 
 ## [1.10.3] - 2026-02-14
 ### Changed
@@ -252,6 +253,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1019]: https://github.com/FossifyOrg/Calendar/issues/1019
 [#1024]: https://github.com/FossifyOrg/Calendar/issues/1024
 [#1065]: https://github.com/FossifyOrg/Calendar/issues/1065
+[#1092]: https://github.com/FossifyOrg/Calendar/issues/1092
 [#1157]: https://github.com/FossifyOrg/Calendar/issues/1157
 
 [Unreleased]: https://github.com/FossifyOrg/Calendar/compare/1.10.3...HEAD